+++ packages/website/components/account/filesManager/fileRowItem.js
@@ -1,12 +1,13 @@
 import clsx from 'clsx';
-import { useCallback, useMemo, useRef, useState } from 'react';
+import { useMemo, useRef } from 'react';
+import { renderToString } from 'react-dom/server';
+import { BsFillInfoCircleFill } from 'react-icons/bs';
 
-import Tooltip from 'modules/zero/components/tooltip/tooltip';
-import CheckIcon from 'assets/icons/check';
-import InfoBIcon from 'assets/icons/infoB';
-import CopyIcon from 'assets/icons/copy';
-import PencilIcon from 'assets/icons/pencil';
-import { addTextToClipboard, truncateString } from 'lib/utils';
+import CheckIcon from '../../../assets/icons/check';
+import CopyIcon from '../../../assets/icons/copy';
+import PencilIcon from '../../../assets/icons/pencil';
+import { addTextToClipboard, formatTimestamp, formatTimestampFull, truncateString } from '../../../lib/utils';
+import Tooltip from '../../../modules/zero/components/tooltip/tooltip';
 import AppData from '../../../content/pages/app/account.json';
 
 export const PinStatus = {
@@ -25,15 +26,16 @@ export const PinStatus = {
  * @property {string} status
  * @property {string} size
  * @property {string | import('react').ReactNode[] | null} storageProviders
- * @property {((e: any)=>void)} onSelect
+ * @property {(e: any)=>void} onSelect
+ * @property {(e: any)=>void} onSelect
  * @property {number} [numberOfPins]
  * @property {boolean} [isHeader]
  * @property {boolean} [isSelected]
  * @property {{text: string, target: "name" | "cid"}} [highlight]
  * @property {()=>void} [onDelete]
- * @property {((newFileName?: string) => void)} [onEditToggle]
+ * @property {(newFileName?: string) => void} [onEditToggle]
  * @property {boolean} [isEditingName]
- * @property {string} tabType
+ * @property {string?} tabType
  */
 
 /**
@@ -70,7 +72,6 @@ const FileRowItem = props => {
     return propsReturn;
   }, [props]);
 
-  const [showAll, setShowAll] = useState(false);
   const fileRowLabels = AppData.page_content.file_manager.table.file_row_labels;
   const statusMessages = fileRowLabels.status.tooltip;
   /** @type {import('react').RefObject<HTMLTextAreaElement>} */
@@ -87,12 +88,15 @@ const FileRowItem = props => {
     [numberOfPins, status, statusMessages]
   );
 
-  const showAllToggle = useCallback(() => {
-    setShowAll(v => !v);
-  }, []);
-
   return (
-    <div className={clsx('files-manager-row', className, isHeader && 'files-manager-row-header')}>
+    <div
+      className={clsx(
+        'files-manager-row',
+        className,
+        isHeader && 'files-manager-row-header',
+        isSelected && 'files-manager-row-active'
+      )}
+    >
       <span className={clsx('file-select-container', tabType)}>
         <span className="file-select">
           <input checked={isSelected} type="checkbox" id={`${name}-select`} onChange={onSelect} />
@@ -101,10 +105,6 @@ const FileRowItem = props => {
         <button onClick={onDelete} className="file-row-label delete medium-down-only">
           {fileRowLabels.delete.label}
         </button>
-      </span>
-      <span className="file-date">
-        <span className="file-row-label medium-down-only">{fileRowLabels.date.label}</span>
-        {date}
       </span>
       <span className={clsx(isEditingName && 'isEditingName', 'file-name')}>
         <span className="file-row-label medium-down-only">{fileRowLabels.name.label}</span>
@@ -125,8 +125,8 @@ const FileRowItem = props => {
       </span>
       <span className="file-cid" title={cid}>
         <span className="file-row-label medium-down-only">
-          <Tooltip content={fileRowLabels.cid.tooltip} />
           {fileRowLabels.cid.label}
+          <Tooltip content={fileRowLabels.cid.tooltip} />
         </span>
         {isHeader ? (
           <span className="cid-full medium-up-only">{cid}</span>
@@ -158,21 +158,21 @@ const FileRowItem = props => {
       </span> */}
       <span className="file-pin-status">
         <span className="file-row-label medium-down-only">
-          <Tooltip content={statusMessages.header} />
           {fileRowLabels.status.label}
+          <Tooltip content={statusMessages.header} />
         </span>
         {status}
         {isHeader ? (
           <Tooltip content={statusMessages.header} />
         ) : (
-          statusTooltip && <Tooltip icon={<InfoBIcon />} content={statusTooltip} />
+          statusTooltip && <Tooltip icon={<BsFillInfoCircleFill />} content={statusTooltip} />
         )}
       </span>
-      {storageProviders ? (
+      {storageProviders && (
         <span className="file-storage-providers">
           <span className="file-row-label medium-down-only">
-            <Tooltip content={fileRowLabels.storage_providers.tooltip.header} />
             {fileRowLabels.storage_providers.label}
+            <Tooltip content={fileRowLabels.storage_providers.tooltip.header} />
           </span>
           {isHeader ? (
             <>
@@ -200,20 +200,35 @@ const FileRowItem = props => {
               />
             </>
           ) : (
-            <div className={clsx('file-storage-providers-content', showAll ? '' : 'show-all')}>
-              <div className="content">{storageProviders}</div>
-              {storageProviders.length > fileRowLabels.storage_providers.toggle_threshold && (
-                <button className="medium-up-only" onClick={showAllToggle}>
-                  {showAll ? 'View fewer' : 'View all'}
-                </button>
-              )}
-            </div>
+            <>
+              Stored ({storageProviders.length})
+              <Tooltip
+                className="medium-down-only"
+                position="left"
+                content={renderToString(<p>{storageProviders}</p>)}
+              />
+              <Tooltip
+                className="medium-up-only"
+                position="right"
+                content={renderToString(<p>{storageProviders}</p>)}
+              />
+            </>
           )}
         </span>
-      ) : null}
+      )}
       <span className="file-size">
         <span className="file-row-label medium-down-only">{fileRowLabels.size.label}</span>
         {size}
+      </span>
+      <span className="file-date">
+        {isHeader ? (
+          <span>Date</span>
+        ) : (
+          <>
+            <span className="file-row-label medium-down-only">{fileRowLabels.date.label}</span>
+            <span title={formatTimestampFull(date)}>{formatTimestamp(date)}</span>
+          </>
+        )}
       </span>
     </div>
   );